+++ src/containers/MainRouter.container.jsx
@@ -2,18 +2,18 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { withRouter } from 'react-router';
 
-import Contribute from './contribute/Contribute.container';
+import Review from './review/Review.container';
 import MainLayout from '../layouts/MainLayout.layout';
 import NotFound from '../components/NotFound.component';
 import Search from './search/Search.container';
-import Resources from './resources/Resources.container';
+import Proposals from './proposals/Proposals.container';
 
 const MainRouter = () => (
     <MainLayout>
         <Switch>
             <Route path="/" exact component={Search} />
-            <Route path="/contribute" exact component={Contribute} />
-            <Route path="/resources" exact component={Resources} />
+            <Route path="/review" exact component={Review} />
+            <Route path="/proposals" exact component={Proposals} />
             <Route path="/notifications" exact component={NotFound} />
             <Route path="/profile" exact component={NotFound} />
             <Route path="/settings" exact component={NotFound} />
