+++ src/containers/proposals/Proposals.container.jsx
@@ -3,22 +3,22 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { setBgClassName } from '../../actions/app.actions';
 
-import styles from './Resources.module.scss';
+import styles from './Proposals.module.scss';
 
-const Resources = (props) => {
+const Proposals = (props) => {
     useEffect(() => {
         props.setBgClassName('bgBlue');
     }, []);
 
     return (
         <div>
-            Resources
+            Proposals
         </div>
     );
 };
 
-Resources.propTypes = {
+Proposals.propTypes = {
     setBgClassName: PropTypes.func,
 };
 
-export default connect(null, { setBgClassName })(Resources);
+export default connect(null, { setBgClassName })(Proposals);
