+++ src/containers/proposals/Proposals.module.scss
